mangastore/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator